requirements.txt
altgraph==0.17.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.59.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
kiwisolver==1.4.7
macholib==1.16.3
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
scipy==1.13.1
shiboken6==6.9.1
six==1.17.0
zipp==3.23.0
